+++ backend/models/user_card.py
@@ -9,7 +9,7 @@ class UserCard(db.Model):
         autoincrement=True)
 
     user_id = db.Column(
-        db.SmallInteger,
+        db.Integer,
         db.ForeignKey('user.id'),
         unique=True,
         nullable=False,
